scanner InterrogatorCLI

+++ src/main/java/com/rendaxx/collection_object/Coordinates.java
@@ -1,6 +1,6 @@
 package com.rendaxx.collection_object;
 
-public record Coordinates(Float x, Double y) {
+public record Coordinates(Double x, Double y) {
     @Override
     public String toString() {
         return "Coordinates{" +
